// src/types/boosts.ts
export interface BoostsResponse {
  albinoLevel: number;
  cobaltAmulet: boolean;
  goldAmulet: boolean;
  hyperMode: boolean;
  potd: boolean;
  seiPower: number;
  shinyChainCount: number;
  shinyChainForme: string;
  shinyChainName: string;
  shinyCharm: boolean;
  silverAmulet: boolean;
  typeRace: boolean;
  uberCharm: boolean;
  zCrystal: boolean;
}